css/css-overscroll-behavior/parsing/overscroll-behavior-x-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overscroll Behavior: parsing overscroll-behavior-x with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-overscroll-behavior-x/#propdef-overscroll-behavior-x">
<meta name="assert" content="overscroll-behavior-x supports the full grammar 'contain | none | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("overscroll-behavior-x", "contain");
test_valid_value("overscroll-behavior-x", "none");
test_valid_value("overscroll-behavior-x", "auto");
</script>
</body>
</html>
